% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner_cmprsk_cmprsk_crr.R
\name{mlr_learners_cmprsk.crr}
\alias{mlr_learners_cmprsk.crr}
\alias{LearnerCompRisksFineGrayCRR}
\title{Fine-Gray Competing Risks Regression Learner}
\description{
A learner for Fine-Gray competing risks regression to estimate cumulative incidence functions (CIFs)
for multiple mutually exclusive events. Calls \code{\link[cmprsk:crr]{cmprsk::crr()}} from \CRANpkg{cmprsk} for training and \code{\link[cmprsk:predict.crr]{cmprsk::predict.crr()}}
for prediction. Supports fixed and time-varying covariates.
}
\section{Dictionary}{

This \link[mlr3:Learner]{Learner} can be instantiated via \link[mlr3:mlr_sugar]{lrn()}:

\if{html}{\out{<div class="sourceCode">}}\preformatted{lrn("cmprsk.crr")
}\if{html}{\out{</div>}}
}

\section{Meta Information}{

\itemize{
\item Task type: \dQuote{cmprsk}
\item Predict Types: \dQuote{cif}
\item Feature Types: \dQuote{logical}, \dQuote{integer}, \dQuote{numeric}, \dQuote{factor}
\item Required Packages: \CRANpkg{mlr3}, \CRANpkg{mlr3proba}, \CRANpkg{cmprsk}
}
}

\section{Parameters}{
\tabular{llll}{
   Id \tab Type \tab Default \tab Range \cr
   maxiter \tab integer \tab 10 \tab \eqn{[1, 1000]}{[1, 1000]} \cr
   gtol \tab numeric \tab 1e-06 \tab \eqn{[1e-09, 0.001]}{[1e-09, 0.001]} \cr
   na.action \tab untyped \tab stats::na.omit \tab - \cr
   init_list \tab untyped \tab NULL \tab - \cr
   censor_group \tab untyped \tab NULL \tab - \cr
   cov2_info \tab untyped \tab NULL \tab - \cr
}
}

\section{Parameter Details}{

\describe{
\item{maxiter}{(\code{integer(1)})\cr Maximum number of iterations for convergence.}
\item{gtol}{(\code{numeric(1)})\cr Convergence tolerance for gradient.}
\item{init_list}{(\code{list()})\cr Initial values for model coefficients, named by event type.}
\item{censor_group}{(\code{character(0)})\cr Censoring group column name, if any.}
\item{cov2_info}{(\code{list()})\cr Information for time-varying covariates, including \code{feat2} (names) and \code{tf} (transformation function).}
}
}

\section{Custom mlr3 parameters}{

\itemize{
\item \code{cov2_info}: Configuration for time-varying features (default: \code{NULL} for model with fixed covariates).
\item \code{init_list}: Initial regression parameters per cause (default: \code{NULL} for zeros).
\item \code{censor_group}: Column name for groups with different censoring distributions (default: \code{character(0)}).
}
}

\section{Initial parameter values}{

\itemize{
\item \code{maxiter}: Set to 100 (package default is 10).
\item \code{gtol}: Set to 1e-6 (package default is 1e-3).
\item \code{na.action}: Set to na.omit.
\item \code{init_list}: Set to NULL (no corresponding parameter in cmprsk
\item \code{censor_group}: Set to NULL,
\item \code{cov2_info}: Set to NULL
}
}

\examples{
\dontshow{if (learner_is_runnable("cmprsk.crr")) withAutoprint(\{ # examplesIf}
library(mlr3)
library(mlr3proba)

# Define the learner
learner = lrn("cmprsk.crr")
print(learner)

# Define a task
task = tsk("pbc")
task$select(c("age", "bili", "sex"))
task$set_col_roles(cols = "status", add_to = "stratum")

# Create train and test sets
ids = partition(task)

# Train the learner
learner$train(task, row_ids = ids$train)

# Print model, convergence, and importance
print(learner$model)
print(learner$convergence())

# Make predictions
predictions = learner$predict(task, ids$test)

# Score the predictions
predictions$score()
\dontshow{\}) # examplesIf}
}
\references{
Fine, P. J, Gray, J. R (1999).
\dQuote{A proportional hazards model for the subdistribution of a competing risk.}
\emph{Journal of the American Statistical Association}, \bold{94}(446), 496--509.
}
\seealso{
\itemize{
\item \link[mlr3misc:Dictionary]{Dictionary} of \link[mlr3:Learner]{Learners}: \link[mlr3:mlr_learners]{mlr3::mlr_learners}.
\item \code{as.data.table(mlr_learners)} for a table of available \link[mlr3:Learner]{Learners} in the running session (depending on the loaded packages).
\item Chapter in the \href{https://mlr3book.mlr-org.com/}{mlr3book}: \url{https://mlr3book.mlr-org.com/basics.html#learners}
\item \CRANpkg{mlr3learners} for a selection of recommended learners.
\item \CRANpkg{mlr3cluster} for unsupervised clustering learners.
\item \CRANpkg{mlr3pipelines} to combine learners with pre- and postprocessing steps.
\item \CRANpkg{mlr3tuning} for tuning of hyperparameters, \CRANpkg{mlr3tuningspaces} for established default tuning spaces.
}
}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[mlr3proba:LearnerCompRisks]{mlr3proba::LearnerCompRisks}} -> \code{LearnerCompRisksFineGrayCRR}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerCompRisksFineGrayCRR-new}{\code{LearnerCompRisksFineGrayCRR$new()}}
\item \href{#method-LearnerCompRisksFineGrayCRR-clone}{\code{LearnerCompRisksFineGrayCRR$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="configure"><a href='../../mlr3/html/Learner.html#method-Learner-configure'><code>mlr3::Learner$configure()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="encapsulate"><a href='../../mlr3/html/Learner.html#method-Learner-encapsulate'><code>mlr3::Learner$encapsulate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="selected_features"><a href='../../mlr3/html/Learner.html#method-Learner-selected_features'><code>mlr3::Learner$selected_features()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerCompRisksFineGrayCRR-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerCompRisksFineGrayCRR-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerCompRisksFineGrayCRR$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerCompRisksFineGrayCRR-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerCompRisksFineGrayCRR-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerCompRisksFineGrayCRR$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
